+++ templates/ruby/example-blank/handler.rb
@@ -1,3 +1,5 @@
+# To deploy this function, cd into its folder and run:
+# faastruby deploy-to WORKSPACE_NAME
 def handler event
   # FUNCTION RESPONSE
   #